Webpack4 - babel - load any babel dependencies immediately, while in the right CWD
contrary to webpack plugins which can be customized via resolveLoader,
babel plugins have no such facility

thus, when passing string names of plugins to babel, there is a current-repo dependency

if babel attempts to resolve such a plugin when running in the context of a ui plugin such as v2v or graphql, it will try to do so from that ui plugin's node_modules

if babel attempts to resolve it when running in ui-classic, this will work

for some reason, `bin/webpack` always loads these from ui-classic, but `node_modules/.bin/webpack` (+params) fails in graphql (the only engine which doesn't have a copy of the plugins now)

Thus, changing this to explicitly `require` every plugin, instead of passing just the names.
That way, all the plugins are loaded when in ui-classic, and won't fail to load afterwards. (In fact, babel won't attempt to load them again.)