Use .count, not .length, for better performance
`.length` on an `ActiveRecord::Relation` object will trigger
`ActiveRecord::Base#load`, which will execute the current query, and
then do a count on that array-like object.

`.count` will make use of the database adapter's SQL `COUNT` call, and
wrap the current `SELECT` statement in a `COUNT(...)`, which will only
return the integer value for the number records that would have been
returned if that query was otherwise run without the `COUNT`.

Since in this case, we are simply using this as an aggregation for a
summary view, the `.count` method is vastly prefered over the `.length`
as doing `.length` will instantiate all the `ActiveRecord` objects in
memory for all of the records before counting the records in ruby.
Doing the same with `.count` not only is faster in ruby, but it is less
strain on the DB as well, since it won't have to transmit almost any
data over the wire as well.